Add dep on fpc-srpm-macros.
Note: I had to either bump the release and not the version, or use a
non-integer version, in order to make sure that this package doesn't end
up newer than the F24 version, which would happen after one more update
which I do intend to make.

Sorry.  The request to keep the version in that manner is simply not
workable in the general case.